+++ app/services/api/v3/database_validation/chain_builders/download_attribute_chain_builder.rb
@@ -1,0 +1,27 @@
+# Prepares checks to be run on download_attribute objects
+
+# The following checks are included:
+#   DONE check years match data in flows
+#   check for widows (download_attributes without download_qual or download_quant)
+module Api
+  module V3
+    module DatabaseValidation
+      module ChainBuilders
+        class DownloadAttributeChainBuilder < AbstractChainBuilder
+          checks :declared_years_match_data,
+                 association: :download_quant,
+                 link: {
+                   method: :admin_download_attribute_path,
+                   params: [:download_attribute]
+                 }
+          checks :declared_years_match_data,
+                 association: :download_qual,
+                 link: {
+                   method: :admin_download_attribute_path,
+                   params: [:download_attribute]
+                 }
+        end
+      end
+    end
+  end
+end
