+++ app/services/api/v3/database_validation/chain_builders/map_attribute_chain_builder.rb
@@ -1,0 +1,27 @@
+# Prepares checks to be run on map_attribute objects
+
+# The following checks are included:
+#   DONE check years match data in flows
+#   check for widows (map_attributes without map_ind or map_quant)
+module Api
+  module V3
+    module DatabaseValidation
+      module ChainBuilders
+        class MapAttributeChainBuilder < AbstractChainBuilder
+          checks :declared_years_match_data,
+                 association: :map_ind,
+                 link: {
+                   method: :admin_map_attribute_path,
+                   params: [:map_attribute]
+                 }
+          checks :declared_years_match_data,
+                 association: :map_quant,
+                 link: {
+                   method: :admin_map_attribute_path,
+                   params: [:map_attribute]
+                 }
+        end
+      end
+    end
+  end
+end
