+++ app/services/api/v3/database_validation/report.rb
@@ -6,7 +6,7 @@
 #   check tooltip text present (WARN)
 module Api
   module V3
-    module DatabaseChecks
+    module DatabaseValidation
       class Report
         def call
           @errors_list = ErrorsList.new
@@ -21,29 +21,29 @@ module Api
         def chain
           chain = []
           Api::V3::Context.all.each do |context|
-            chain += ContextCheckChain.new(context, @errors_list).chain
+            chain += ChainBuilders::ContextChainBuilder.new(context, @errors_list).chain
             context.context_node_types.each do |context_node_type|
-              chain += ContextNodeTypeCheckChain.new(
+              chain += ChainBuilders::ContextNodeTypeChainBuilder.new(
                 context_node_type, @errors_list
               ).chain
             end
             context.resize_by_attributes.each do |resize_by_attribute|
-              chain += ResizeByAttributeCheckChain.new(
+              chain += ChainBuilders::ResizeByAttributeChainBuilder.new(
                 resize_by_attribute, @errors_list
               ).chain
             end
             context.recolor_by_attributes.each do |recolor_by_attribute|
-              chain += RecolorByAttributeCheckChain.new(
+              chain += ChainBuilders::RecolorByAttributeChainBuilder.new(
                 recolor_by_attribute, @errors_list
               ).chain
             end
             context.download_attributes.each do |download_attribute|
-              chain += DownloadAttributeCheckChain.new(
+              chain += ChainBuilders::DownloadAttributeChainBuilder.new(
                 download_attribute, @errors_list
               ).chain
             end
             context.map_attributes.each do |map_attribute|
-              chain += MapAttributeCheckChain.new(
+              chain += ChainBuilders::MapAttributeChainBuilder.new(
                 map_attribute, @errors_list
               ).chain
             end
