+++ spec/services/api/v3/database_validation/checks/declared_years_match_data_spec.rb
@@ -1,7 +1,7 @@
 require 'rails_helper'
-require 'modules/api/v3/database_checks/shared_check_examples'
+require 'services/api/v3/database_validation/checks/shared_check_examples'
 
-RSpec.describe Api::V3::DatabaseChecks::DeclaredYearsMatchData do
+RSpec.describe Api::V3::DatabaseValidation::Checks::DeclaredYearsMatchData do
   context 'when checking resize_by_attributes' do
     before do
       Api::V3::ResizeByAttribute.skip_callback(:commit, :after, :refresh_dependencies)
@@ -34,13 +34,13 @@ RSpec.describe Api::V3::DatabaseChecks::DeclaredYearsMatchData do
     }
 
     let(:check) {
-      Api::V3::DatabaseChecks::DeclaredYearsMatchData.new(
+      Api::V3::DatabaseValidation::Checks::DeclaredYearsMatchData.new(
         resize_by_attribute,
         association: :resize_by_quant
       )
     }
     let(:report_status) {
-      Api::V3::DatabaseChecks::ErrorsList.new
+      Api::V3::DatabaseValidation::ErrorsList.new
     }
     context 'when years dont match' do
       include_examples 'failing checks'
